Restructured the setup of utils without dependencies of dl_framework
model_summary is now in hooks.py, because it needs hooks. there are no
more dependencies of dl_framework in utils
