+++ dl_framework/hooks.py
@@ -1,5 +1,5 @@
 from functools import partial
-from dl_framework.utils import listify
+from dl_framework.utils import ListContainer, get_batch, find_modules, is_lin_layer
 from torch import Tensor
 
 
@@ -14,37 +14,37 @@ class Hook():
         self.remove()
 
 
-class ListContainer():
-    def __init__(self, items):
-        self.items = listify(items)
+# class ListContainer():
+#     def __init__(self, items):
+#         self.items = listify(items)
 
-    def __getitem__(self, idx):
-        try:
-            return self.items[idx]
-        except TypeError:
-            if isinstance(idx[0], bool):
-                assert len(idx) == len(self)  # bool mask
-                return [o for m, o in zip(idx, self.items) if m]
-            return [self.items[i] for i in idx]
+#     def __getitem__(self, idx):
+#         try:
+#             return self.items[idx]
+#         except TypeError:
+#             if isinstance(idx[0], bool):
+#                 assert len(idx) == len(self)  # bool mask
+#                 return [o for m, o in zip(idx, self.items) if m]
+#             return [self.items[i] for i in idx]
 
-    def __len__(self):
-        return len(self.items)
+#     def __len__(self):
+#         return len(self.items)
 
-    def __iter__(self):
-        return iter(self.items)
+#     def __iter__(self):
+#         return iter(self.items)
 
-    def __setitem__(self, i, o):
-        self.items[i] = o
+#     def __setitem__(self, i, o):
+#         self.items[i] = o
 
-    def __delitem__(self, i):
-        del(self.items[i])
+#     def __delitem__(self, i):
+#         del(self.items[i])
 
-    def __repr__(self):
-        res = f'{self.__class__.__name__} \
-                ({len(self)} items)\n{self.items[:10]}'
-        if len(self) > 10:
-            res = res[:-1] + '...]'
-        return res
+#     def __repr__(self):
+#         res = f'{self.__class__.__name__} \
+#                 ({len(self)} items)\n{self.items[:10]}'
+#         if len(self) > 10:
+#             res = res[:-1] + '...]'
+#         return res
 
 
 class Hooks(ListContainer):
@@ -67,6 +67,23 @@ class Hooks(ListContainer):
     def remove(self):
         for h in self:
             h.remove()
+
+
+# aus utils hier rüberkopiert, weil hooks benötigt werden
+def model_summary(run, learn, data, find_all=False):
+    xb, yb = get_batch(data.valid_dl, run)
+    # Model may not be on the GPU yet
+    device = next(learn.model.parameters()).device
+    xb, yb = xb.to(device), yb.to(device)
+    if find_all:
+        mods = find_modules(learn.model, is_lin_layer)
+    else:
+        learn.model.children()
+    # mods = find_modules(learn.model, is_lin_layer)
+    # if find_all else learn.model.children()
+    f = lambda hook, mod, inp, out: print(f"{mod}\n{out.shape}\n")
+    with Hooks(mods, f) as hooks:
+        learn.model(xb)
 
 # ab hier aus dem source code von fastai kopiert
 # https://github.com/fastai/fastai/blob/master/fastai/callbacks/hooks.py#L58
