+++ dl_framework/utils.py
@@ -1,7 +1,6 @@
 import re
 from typing import Iterable
 from torch import nn
-from dl_framework.hooks import Hooks
 
 
 def lin_comb(v1, v2, beta): return beta*v1 + (1-beta)*v2
@@ -111,39 +110,34 @@ def compose(x, funcs, *args, order_key='_order', **kwargs):
     return x
 
 
-def model_summary(run, learn, data, find_all=False):
-    xb, yb = get_batch(data.valid_dl, run)
-    # Model may not be on the GPU yet
-    device = next(learn.model.parameters()).device
-    xb, yb = xb.to(device), yb.to(device)
-    mods = find_modules(learn.model, is_lin_layer) if find_all else learn.model.children()
-    f = lambda hook, mod, inp, out: print(f"{mod}\n{out.shape}\n")
-    with Hooks(mods, f) as hooks:
-        learn.model(xb)
-
-
 class ListContainer():
-    def __init__(self, items): self.items = listify(items)
+    def __init__(self, items):
+        self.items = listify(items)
 
     def __getitem__(self, idx):
         try:
             return self.items[idx]
         except TypeError:
             if isinstance(idx[0], bool):
-                assert len(idx) == len(self)    # bool mask
+                assert len(idx) == len(self)  # bool mask
                 return [o for m, o in zip(idx, self.items) if m]
             return [self.items[i] for i in idx]
 
-    def __len__(self): return len(self.items)
+    def __len__(self):
+        return len(self.items)
 
-    def __iter__(self): return iter(self.items)
+    def __iter__(self):
+        return iter(self.items)
 
-    def __setitem__(self, i, o): self.items[i] = o
+    def __setitem__(self, i, o):
+        self.items[i] = o
 
-    def __delitem__(self, i): del(self.items[i])
+    def __delitem__(self, i):
+        del(self.items[i])
 
     def __repr__(self):
-        res = f'{self.__class__.__name__} ({len(self)} items)\n{self.items[:10]}'
+        res = f'{self.__class__.__name__} \
+                ({len(self)} items)\n{self.items[:10]}'
         if len(self) > 10:
             res = res[:-1] + '...]'
         return res
